Update custom-ingress-configuration.adoc
The phrase "In this case, you must apply a custom ConfigMap using the template." lead an actual customer to assume that it's always possible to configure an ingress using a ConfigMap, but that is not true. They tried to do path-based routing with NGINX Inc ingress controller and it's not possible. As this can be triggered with rtfctl, looks like it's part of our responsibility to solve everything with a ConfigMap, but it's not possible. The idea is to reframe and warn that ingress is a responsibility of the customer on the Self-Managed Kubernetes solution, at least for now.